face-attendance/requirements.txt
opencv-python==4.8.1.78
deepface==0.0.79
flask==2.3.3
pandas==2.1.1
numpy==1.24.3
tensorflow==2.13.0
pillow==10.0.1
